Content_Management_System/requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
autopep8==1.7.0
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
constantly==15.1.0
cryptography==38.0.1
daphne==3.0.2
dj-database-url==1.2.0
Django==4.1.1
django-appconf==1.0.5
django-cors-headers==3.14.0
django-crispy-forms==1.14.0
django-filter==22.1
django-heroku==0.3.1
django-imagekit==2.0.4
django-schema-graph==2.1.0
django-silk==0.1
djangorestframework==3.14.0
firebase-admin==5.4.0
geographiclib==2.0
geopy==2.3.0
google-api-core==2.10.1
google-api-python-client==2.63.0
google-auth==2.12.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.7.0
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.49.1
grpcio-status==1.49.1
httplib2==0.20.4
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
Jinja2==3.1.2
MarkupSafe==2.1.1
msgpack==1.0.4
Pillow==9.2.0
proto-plus==1.22.1
protobuf==4.21.7
psycopg2==2.9.5
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.13.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
rest-firebase-auth==0.0.2
rsa==4.9
sentry-sdk==1.9.9
service-identity==21.1.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.3
toml==0.10.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.4
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.3.0
zope.interface==5.4.0
